backend/requirements.txt
Django==4.2.12
django-appconf==1.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
pillow==10.4.0
pytest==8.3.3
pytest-cov==5.0.0
pytest-django==4.8.0
